% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calculate_predictions_helper}
\alias{calculate_predictions_helper}
\title{Internal helper to calculate predictions for gof}
\usage{
calculate_predictions_helper(model, data)
}
\arguments{
\item{model}{A redeem_result object.}

\item{data}{A data.table containing the covariates.}
}
\description{
Internal helper to calculate predictions for gof
}
\keyword{internal}
