% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_terms.R
\name{check.RedeemTerm}
\alias{check.RedeemTerm}
\title{Check Arguments for redeem Model Terms}
\usage{
check.RedeemTerm(
  arglist,
  expected = list(),
  defaults = list(),
  allowed_processes = c("0-1", "1-0"),
  allowed_models = c("dem", "rem"),
  model_type = "dem",
  directed = NULL,
  directed_only = FALSE,
  undirected_only = FALSE
)
}
\arguments{
\item{arglist}{List of arguments.}

\item{expected}{List of expected types/values.}

\item{defaults}{List of default values.}

\item{allowed_processes}{Vector of allowed processes ("0-1", "1-0").}

\item{model_type}{Current model type.}
}
\description{
Internal helper to validate arguments and set defaults.
}
\keyword{internal}
