% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{dem_bipartite}
\alias{dem_bipartite}
\title{Bipartite Relational and Durational Event Model (DEM)}
\usage{
dem_bipartite(
  events,
  training_start = 0,
  exogenous_end = NULL,
  formula_0_1 = NULL,
  formula_1_0 = NULL,
  n_nodes = NULL,
  directed = TRUE,
  estimate_0_1 = NULL,
  estimate_1_0 = NULL,
  coef_0_1 = NULL,
  coef_1_0 = NULL,
  simultaneous_interactions = TRUE,
  control = control.redeem()
)
}
\arguments{
\item{events}{The event matrix.}

\item{training_start}{Start time.}

\item{exogenous_end}{End time.}

\item{formula_0_1}{Formula for start events.}

\item{formula_1_0}{Formula for end events.}

\item{n_nodes}{Senders and receivers counts.}

\item{directed}{Directed flag.}

\item{estimate_0_1}{Initial 0-1 estimate.}

\item{estimate_1_0}{Initial 1-0 estimate.}

\item{coef_0_1}{Initial 0-1 coefficients.}

\item{coef_1_0}{Initial 1-0 coefficients.}

\item{simultaneous_interactions}{Simultaneous interactions flag.}

\item{control}{Control parameters.}
}
\description{
Bipartite Relational and Durational Event Model (DEM)
}
