% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{dem_bipartite_simulate}
\alias{dem_bipartite_simulate}
\title{Simulate Bipartite DEM}
\usage{
dem_bipartite_simulate(
  n_nodes,
  coef_0_1,
  coef_1_0,
  degree_coef_0_1 = NULL,
  degree_coef_1_0 = NULL,
  n_events,
  simultaneous_interactions = TRUE,
  seed = 123
)
}
\arguments{
\item{n_nodes}{Senders and receivers counts.}

\item{coef_0_1}{Start coefficients.}

\item{coef_1_0}{End coefficients.}

\item{degree_coef_0_1}{Start degree coefficients.}

\item{degree_coef_1_0}{End degree coefficients.}

\item{n_events}{Number of events.}

\item{simultaneous_interactions}{Simultaneous interactions flag.}

\item{seed}{Random seed.}
}
\description{
Simulate Bipartite DEM
}
