% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{formula_preprocess_single}
\alias{formula_preprocess_single}
\title{Preprocess a single formula for model terms}
\usage{
formula_preprocess_single(
  formula,
  n_nodes,
  model_type = "dem",
  process = "0-1",
  directed = FALSE,
  simulation = FALSE
)
}
\arguments{
\item{formula}{An R formula object.}

\item{n_nodes}{Number of nodes.}

\item{model_type}{Either "dem" or "rem".}

\item{process}{Either "0-1" (incidence) or "1-0" (duration).}

\item{directed}{Logical; if TRUE, the model is directed (defaults to FALSE).}

\item{simulation}{Logical; if TRUE, the formula is being preprocessed for simulation (defaults to FALSE).}
}
\description{
This function takes an R formula and extracts the necessary information to build the model's
design matrix. It identifies special terms, transformations, and associated data.
}
\keyword{internal}
