% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{gof_bipartite}
\alias{gof_bipartite}
\title{Goodness of Fit for Bipartite Models}
\usage{
gof_bipartite(
  object,
  edgelist_test,
  edgelist_train = NULL,
  k_max = 10,
  ties.method = "average"
)
}
\arguments{
\item{object}{The fitted model object.}

\item{edgelist_test}{Test edgelist.}

\item{edgelist_train}{Train edgelist.}

\item{k_max}{Maximum cutpoint.}

\item{ties.method}{Ties method.}
}
\description{
Goodness of Fit for Bipartite Models
}
