% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_terms.R
\name{InitRedeemTerm}
\alias{InitRedeemTerm}
\title{Initialize redeem Model Terms}
\usage{
InitRedeemTerm(
  term_name,
  arglist,
  model_type,
  process,
  n_nodes,
  directed = FALSE
)
}
\arguments{
\item{term_name}{The name of the term.}

\item{arglist}{A list of arguments passed to the term in the formula.}

\item{model_type}{Either "dem" or "rem".}

\item{process}{Either "0-1" (incidence) or "1-0" (duration).}

\item{n_nodes}{Number of nodes in the network.}

\item{directed}{Logical; if TRUE, the model is directed. Used for
term-specific logic.}
}
\description{
This is an internal dispatcher that calls the appropriate term
initialization function.
}
\keyword{internal}
