% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{plot_baseline}
\alias{plot_baseline}
\title{Plot Baseline Intensity}
\usage{
plot_baseline(x, ...)
}
\arguments{
\item{x}{A \code{\link{dem}}, \code{\link{rem}}, or
\code{\link[=summary.redeem_result]{redeem_result}} object.}

\item{...}{Additional arguments passed to `graphics::plot`.}
}
\description{
Plot Baseline Intensity
}
