% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{predict_transition_helper}
\alias{predict_transition_helper}
\title{Helper to predict single transition process}
\usage{
predict_transition_helper(model, data, type)
}
\description{
Helper to predict single transition process
}
\keyword{internal}
