% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.R
\name{prepare_residual_plot_data}
\alias{prepare_residual_plot_data}
\title{Internal helper to prepare residual plot data}
\usage{
prepare_residual_plot_data(raw_list)
}
\arguments{
\item{raw_list}{List of raw residuals}
}
\description{
Internal helper to prepare residual plot data
}
\keyword{internal}
