% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{process_event_actors}
\alias{process_event_actors}
\title{Process event actor columns and automatically identify or validate n_nodes}
\usage{
process_event_actors(events, n_nodes = NULL, directed = TRUE)
}
\arguments{
\item{events}{A matrix or data frame of events with columns \code{from} and \code{to} (or columns 2 and 3).}

\item{n_nodes}{Integer; the total number of actors in the network, or \code{NULL}.}

\item{directed}{Logical; whether the interaction events are directed. Defaults to TRUE.}
}
\value{
A list containing \code{events} (potentially modified) and \code{n_nodes}.
}
\description{
Process event actor columns and automatically identify or validate n_nodes
}
\keyword{internal}
