% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redeem-package.R
\docType{package}
\name{redeem}
\alias{redeem-package}
\alias{redeem}
\title{redeem: Scalable Durational Event Models}
\description{
The \pkg{redeem} package provides a comprehensive framework for the estimation
of Durational Event Models (DEM) and Relational Event Models (REM).
}
\details{
The package is built to handle relational event sequences where interactions
have specific start and end times. It implements the scalable block-coordinate
ascent algorithm described in Fritz et al. (2026), which allows for the
estimation of high-dimensional actor-specific effects and time-varying baseline
intensities.

Key functions:
\itemize{
  \item \code{\link{dem}}: Estimates a Durational Event Model.
  \item \code{\link{rem}}: Estimates a Relational Event Model.
}
}
\references{
Fritz, C., Rastelli, R., Fop, M., & Caimo, A. (2026). Scalable Durational Event Models:
Application to Physical and Digital Interactions. arXiv:2504.00049.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://corneliusfritz.github.io/redeem/}
  \item Report bugs at \url{https://github.com/corneliusfritz/redeem/issues}
}

}
\author{
Cornelius Fritz
}
\keyword{internal}
