% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{rem_bipartite}
\alias{rem_bipartite}
\title{Bipartite Relational Event Model (REM)}
\usage{
rem_bipartite(
  events,
  training_start = 0,
  exogenous_end = NULL,
  formula = NULL,
  n_nodes = NULL,
  directed = TRUE,
  estimate = NULL,
  coef = NULL,
  control = control.redeem()
)
}
\arguments{
\item{events}{The event matrix.}

\item{training_start}{Start time.}

\item{exogenous_end}{End time.}

\item{formula}{Formula for events.}

\item{n_nodes}{Senders and receivers counts.}

\item{directed}{Directed flag.}

\item{estimate}{Initial estimate.}

\item{coef}{Initial coefficients.}

\item{control}{Control parameters.}
}
\description{
Bipartite Relational Event Model (REM)
}
