% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{rem_bipartite_simulate}
\alias{rem_bipartite_simulate}
\title{Simulate Bipartite REM}
\usage{
rem_bipartite_simulate(
  n_nodes,
  coef,
  degree_coef = NULL,
  n_events,
  time_changepoints = NULL,
  baseline = NULL,
  seed = 123
)
}
\arguments{
\item{n_nodes}{Senders and receivers counts.}

\item{coef}{Coefficients.}

\item{degree_coef}{Degree coefficients.}

\item{n_events}{Number of events to simulate.}

\item{time_changepoints}{Time changepoints.}

\item{baseline}{Baseline vector.}

\item{seed}{Random seed.}
}
\description{
Simulate Bipartite REM
}
