% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reproduce_model_data}
\alias{reproduce_model_data}
\title{Reconstruct estimation data from a model object}
\usage{
reproduce_model_data(models, verbose = FALSE)
}
\arguments{
\item{models}{A redeem model object (rem or dem).}

\item{verbose}{Logical.}
}
\description{
Reconstruct estimation data from a model object
}
\keyword{internal}
