% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rhs_terms_as_list}
\alias{rhs_terms_as_list}
\title{Convert right-hand side of a formula to a list of term information}
\usage{
rhs_terms_as_list(
  formula,
  n_nodes,
  env = NULL,
  evaluate_calls = FALSE,
  model_type = "dem",
  process = "0-1",
  directed = FALSE
)
}
\arguments{
\item{formula}{The formula to parse.}

\item{n_nodes}{Number of nodes.}

\item{env}{The environment in which to evaluate terms.}

\item{evaluate_calls}{Logical; if `TRUE`, evaluates the full calls.}

\item{model_type}{Either "dem" or "rem".}

\item{process}{Either "0-1" or "1-0".}

\item{directed}{Logical; if TRUE, the model is directed.}
}
\value{
A list of term information, including labels, base names, and data.
}
\description{
Convert right-hand side of a formula to a list of term information
}
\keyword{internal}
