% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{standardize_redeem_result}
\alias{standardize_redeem_result}
\title{Standardize Estimation Output for Redeem Models}
\usage{
standardize_redeem_result(
  coefficients,
  coef_hist,
  covariance,
  llh_hist,
  data,
  prediction,
  method = "dem.nr",
  n_nodes = NULL,
  directed = FALSE,
  time_changepoints = NULL,
  labels_changepoints = NULL,
  full_baseline = FALSE,
  return_data = TRUE,
  save_hist = TRUE
)
}
\arguments{
\item{coefficients}{Joint coefficient vector.}

\item{coef_hist}{Matrix of coefficient history.}

\item{covariance}{Covariance matrix for core effects.}

\item{llh_hist}{Vector of log-likelihood history.}

\item{data}{Input data.}

\item{prediction}{Predicted values.}

\item{method}{Estimation method name (for class).}

\item{n_nodes}{Number of nodes.}

\item{directed}{Logical; are the interaction events directed?}

\item{time_changepoints}{Numeric vector of time changepoints.}

\item{labels_changepoints}{Character vector of labels for the changepoints.}

\item{return_data}{Logical; should the estimation dataset be returned?}

\item{save_hist}{Logical; should the parameter history be returned?}
}
\description{
Standardize Estimation Output for Redeem Models
}
\keyword{internal}
