% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_models.R
\name{update_core_r}
\alias{update_core_r}
\title{Update core coefficients using R (for correctness assessment)}
\usage{
update_core_r(
  data,
  covarites,
  prediction,
  est_core,
  identifiable,
  offset_fixed = NULL
)
}
\arguments{
\item{data}{Preprocessed data.table.}

\item{prediction}{Current predicted intensities.}

\item{est_core}{Initial core coefficients.}

\item{identifiable}{Logical vector indicating identifiable coefficients.}

\item{offset_fixed}{Fixed offset (degree + baseline).}
}
\description{
Update core coefficients using R (for correctness assessment)
}
\keyword{internal}
