% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{summary.redeem_result}
\alias{summary.redeem_result}
\title{Summary of REDEEM Model Results}
\usage{
\method{summary}{redeem_result}(object, ...)
}
\arguments{
\item{object}{A `redeem_result` object.}

\item{...}{Additional arguments.}
}
\description{
Summary of REDEEM Model Results
}
